feat(expression-compatibility): break `ffi.rs` into module.
